chore: Port Over Login/Reset Pages from Svelte Version
Included Change:
- Added and configured Sass using the Remix Docs

These direct ports from our "svelte-kit-supertokens"
repository are not necessarily "perfect" or "optimal".
Handling styles in Remix land feels a little annoying
(which isn't surprising, given that this is a
React project). We may re-evaluate our approach
in the future.

As for porting everything else, it's probably fine
as it is... HTML, JS, and JSX are all pretty
straightforward, though perhaps there are some
additional improvements we can make.