chore: WIP - Get Working Draft for Logging in with Remix + SuperTokens
Note: We are a far way from being done. It seems it would be
best to get SuperTokens working with Remix _WITHOUT_ the need
for JavaScript. (i.e., we want to try an HTML-only approach)

The code can also be cleaned up better. But we've made so much
progress that we think it might be better to do things
more piece by piece from here on out.