feat: Enable Users to Reset Their Passwords (Draft)
With this new feature, users can successfully request
a password reset via email (and then proceed to reset
their password). NOTE THAT THIS IS HAPPY PATH ONLY!
We need to add the unhappy paths for the password
resets and basically all of our auth stuff.

We're committing what we have now so that we can
move forward more easily without losing our work.

Included Changes:
- Make `domain` an environment variable. ... It seems
  this is just necessary for the server to know the
  proper URLs to fetch.
- Style
     - Move `login.scss` link styles to `auth-form.scss`.
     - Improve `global` button styles.
     - Create styles for SuperTokens "error banner" (appears
       between headers and inputs typically).
     - Add "auth-card" class to style SuperTokens-like
       cards that are not auth forms.