feat: Add Error Messages for Login Forms (SSR Only; Draft)
Included Changes:
- Added (accessible) error messages for login/sign up form.
   - Error messages come directly from SuperTokens (or mimic
     what SuperTokens example apps do in the case of "banners").
   - Attempted to style messages based on what SuperTokens does.
- Corrected Form `action` based on Remix error (see GitHub issue).
- Removed redundant email existence check.
   - Apparently SuperTokens does this already while doing signups.
     So this endpoint is really only useful for the frontend.

NOTE: This is "SSR-ONLY". Users not using JavaScript will
see the errors (thanks to SSR). As for JS users, they will
also see the error. However, JS users _will not_ be able
to dismiss the error by interacting with the inputs yet.
**We still need to add client side logic to dismiss the
error messages if appropriate. Would `react-hook-form`
be useful for this? Do we even need it?**