chore: Remove `supertokens-webiste` and Update Notes
We've decided that using `localStorage` will be the
best approach for preserving user data for complex
forms. (This is because `localStorage` will have to
be used anyway if the user refreshes the page, not just
if they get have to re-authenticate. So approaching
things this way keeps life simple for many reasons.)
Therefore, `supertokens-website` has finally been removed.
We've updated our notes accordingly.

We tried to make the README smaller so that it's
easier to read. Also, now that the repo is public,
our silly comments that are a little more sassy
by nature have been removed to decrease the risk
of negative misinterpretations of our notes.

Split out SuperTokens suggestions into their own file.