refactor: Update Folder Structure for Styles
For the most part, the styles folder should reflect
the folder structure of the `/app` folder. This is not
pretty, but it's more or less the best thing we've got
with Remix... at least for now.

"Global" and "re-used"/"shared" styles are unique,
and they will be placed in the `/shared` folder.

Perhaps we should update this in the README?