refactor: Get `SuperTokens` Working without Middleware
This was basically required in order to use HTTPS
and Cloudflare simultaneously with proper security
and efficiency. See the updated notes in this
commit for additional details.

Note that this approach _should_ also open up the
door for working with other SSR frameworks. We
still need to do some POCs.

**HOWEVER**: This is only intended to be a
TEMPORARY solution. Hopefully the SuperTokens
team will improve `supertokens-node` so that
this slightly-complex code would become
unnecessary. Again, see our updated notes
in this commit.