refactor: Require Header Inputs As a `Map` Instead of `Headers`
This approach is simply more versatile. Not every SSR
framework will have a `Headers` representation, and
Node.js does not yet support `Headers` reliably
locally.

As you can see in the changes, there's very little
additional effort on our end with this refactor (AND
on the developer's end). So really, this is better for
everyone right now.