chore: Remove Unused Route Check from Auth Logic
The `requestAllowed` "Conditional Path" in the
auth logic already guarantees that recursive
redirects will not happen during refreshes.
So the hack that was used earlier is no longer
necessary. (The hack was also limited in its
effectiveness since it _seems_ like it caused
the `returnUrl` to be forgotten during valid
refreshes.)