chore: Upgrade Remix to Latest and Use Vite
This simplifies the effort needed to setup the
server for development. It also enables console
logs on the server side to start working again.

NOTE: We tested the basic pages and signing in/out.
However, we did not test `/auth/session/refresh`.
We're assuming that if the other SuperTokens
features worked, this one should work fine as well.