chore: Derive `useActionData` Type from `action` Server Function
In the case of our example application, this actually doesn't
provide any significant benefit. This is more so for the sake
of example. However, in other kinds of applications, it can be
VERY helpful for the `useActionData` return type to be implicitly
derived from the `action` server function. These changes are
to show other Remix Developers how to go about that.

(Technique was learned from SvelteKit.)