refactor: Run Auth Tests serially to avoid race conditions
In some cases, the `existingAccount` Worker-scoped Fixture would be
executed multiple times in parallel (because the tests were configured
to run in a fully-parallel manner based on `playright.config.ts`).
The SQLite database that SuperTokens uses for its in-memory database
was not able to handle simultaneous user creation, so it would cause
the tests to fail whenever this happened.

Now we run the tests serially so that only 1 Worker will be used for
all the tests in `auth.test.ts`. Consequently, only 1 user account
will be created by the `existingAccount` Worker-scoped Fixture, and
we won't need to worry about the race condition we described earlier.

See: https://github.com/microsoft/playwright/issues/32354