WIP: Add ThirdParty Integration
The code seems to be more or less functional. We just want to clean
up the implementation a little bit to make the code more understandable.
We also want to make sure there aren't any hidden bugs that would
occur when the provider redirects users from the ThirdParty Login Page.